+++ BEPUik/IKPointOnLineJoint.cs
@@ -67,10 +67,10 @@ namespace BEPUik
         {
             Vector3 cross;
             Vector3.Cross(ref localLineDirection, ref Toolbox.UpVector, out cross);
-            float lengthSquared = cross.LengthSquared();
+            double lengthSquared = cross.LengthSquared();
             if (lengthSquared > Toolbox.Epsilon)
             {
-                Vector3.Divide(ref cross, (float)Math.Sqrt(lengthSquared), out localRestrictedAxis1);
+                Vector3.Divide(ref cross, (double)Math.Sqrt(lengthSquared), out localRestrictedAxis1);
             }
             else
             {
@@ -121,7 +121,7 @@ namespace BEPUik
             //Find the point on the line closest to the world point.
             Vector3 offset;
             Vector3.Subtract(ref worldPoint, ref worldLineAnchor, out offset);
-            float distanceAlongAxis;
+            double distanceAlongAxis;
             Vector3.Dot(ref offset, ref lineDirection, out distanceAlongAxis);
 
             Vector3 worldNearPoint;
