+++ BEPUik/IKRevoluteJoint.cs
@@ -68,13 +68,13 @@ namespace BEPUik
         {
             Vector3 worldAxisA = WorldFreeAxisA;
             Vector3 error = Vector3.Cross(worldAxisA, WorldFreeAxisB);
-            float lengthSquared = error.LengthSquared();
+            double lengthSquared = error.LengthSquared();
             Vector3 worldConstrainedAxis1, worldConstrainedAxis2;
             //Find the first constrained axis.
             if (lengthSquared > Toolbox.Epsilon)
             {
                 //The error direction can be used as the first axis!
-                Vector3.Divide(ref error, (float)Math.Sqrt(lengthSquared), out worldConstrainedAxis1);
+                Vector3.Divide(ref error, (double)Math.Sqrt(lengthSquared), out worldConstrainedAxis1);
             }
             else
             {
@@ -85,7 +85,7 @@ namespace BEPUik
                 if (lengthSquared > Toolbox.Epsilon)
                 {
                     //The up vector worked!
-                    Vector3.Divide(ref worldConstrainedAxis1, (float)Math.Sqrt(lengthSquared), out worldConstrainedAxis1);
+                    Vector3.Divide(ref worldConstrainedAxis1, (double)Math.Sqrt(lengthSquared), out worldConstrainedAxis1);
                 }
                 else
                 {
