+++ BEPUutilities/RayHit.cs
@@ -17,6 +17,6 @@
         /// T parameter of the ray hit.  
         /// The ray hit location is equal to the ray origin added to the ray direction multiplied by T.
         ///</summary>
-        public float T;
+        public double T;
     }
 }
